+++ tests/LocalData/LocalDataTest.php
@@ -135,20 +135,15 @@ class LocalDataTest extends TestCase
 
     public function testLocalDataAttributeUnauthorized()
     {
-        // authorization expression of attribute evaluates to false -> value must be null
         $localDataAttributeName = 'attribute_3';
         $sourceData = ['src_attribute_3' => 'value_3'];
         $this->testDataProvider->setSourceData($sourceData);
-        $entity = $this->getTestEntityWithLocalData($localDataAttributeName, $sourceData);
-        $this->assertNull($entity->getLocalDataValue($localDataAttributeName));
-
-        $includeLocal = 'attribute_2,attribute_3';
-        $sourceData = ['src_attribute_3' => 'value_3', 'src_attribute_2' => 'value_2'];
-
-        $this->testDataProvider->setSourceData($sourceData);
-        $entity = $this->getTestEntityWithLocalData($includeLocal, $sourceData);
-        $this->assertNull($entity->getLocalDataValue('attribute_3'));
-        $this->assertEquals('value_2', $entity->getLocalDataValue('attribute_2'));
+        try {
+            $this->getTestEntityWithLocalData($localDataAttributeName, $sourceData);
+            $this->fail('Expected ApiError exception not thrown.');
+        } catch (ApiError $e) {
+            $this->assertEquals(Response::HTTP_FORBIDDEN, $e->getStatusCode());
+        }
     }
 
     public function testLocalDataAttributeUnauthorizedCollection()
@@ -158,29 +153,16 @@ class LocalDataTest extends TestCase
         $sourceData = ['src_attribute_3' => 'value_3'];
 
         $this->testDataProvider->setSourceData($sourceData);
-        $entities = $this->getTestEntitiesWithLocalData($localDataAttributeName, [
-            '0' => $sourceData,
-            '1' => $sourceData,
-            '2' => $sourceData,
-        ]);
-
-        $this->assertNull($entities[0]->getLocalDataValue($localDataAttributeName));
-        $this->assertNull($entities[1]->getLocalDataValue($localDataAttributeName));
-        $this->assertNull($entities[2]->getLocalDataValue($localDataAttributeName));
-
-        $includeLocal = 'attribute_3,attribute_1';
-        $sourceData = ['src_attribute_1' => 'value_1', 'src_attribute_3' => 'value_3'];
-
-        $this->testDataProvider->setSourceData($sourceData);
-        $entities = $this->getTestEntitiesWithLocalData($includeLocal, [
-            '0' => $sourceData,
-            '1' => $sourceData,
-        ]);
-
-        $this->assertNull($entities[0]->getLocalDataValue('attribute_3'));
-        $this->assertEquals('value_1', $entities[0]->getLocalDataValue('attribute_1'));
-        $this->assertNull($entities[1]->getLocalDataValue('attribute_3'));
-        $this->assertEquals('value_1', $entities[1]->getLocalDataValue('attribute_1'));
+        try {
+            $this->getTestEntitiesWithLocalData($localDataAttributeName, [
+                '0' => $sourceData,
+                '1' => $sourceData,
+                '2' => $sourceData,
+            ]);
+            $this->fail('Expected ApiError exception not thrown.');
+        } catch (ApiError $e) {
+            $this->assertEquals(Response::HTTP_FORBIDDEN, $e->getStatusCode());
+        }
     }
 
     /**